Consolidate exceptions in to `exceptions.py`
