+++ src/briefcase/integrations/android_sdk.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from briefcase.config import PEP508_NAME_RE
 from briefcase.console import InputDisabled, select_option
 from briefcase.exceptions import (
+    AndroidDeviceNotAuthorized,
     BriefcaseCommandError,
     InvalidDeviceError,
     MissingToolError,
@@ -18,24 +19,6 @@ from briefcase.integrations.base import Tool, ToolCache
 from briefcase.integrations.java import JDK
 
 DEVICE_NOT_FOUND = re.compile(r"^error: device '[^']*' not found")
-
-
-class AndroidDeviceNotAuthorized(BriefcaseCommandError):
-    def __init__(self, device):
-        self.device = device
-        super().__init__(
-            f"""
-The device you have selected ({device}) has not had developer options and
-USB debugging enabled. These must be enabled before a device can be used  as a
-target for deployment. For details on how to enable Developer Options, visit:
-
-    https://developer.android.com/studio/debug/dev-options#enable
-
-Once you have enabled these options on your device, you will be able to select
-this device as a deployment target.
-
-"""
-        )
 
 
 class AndroidSDK(Tool):
